+++ packages/domain-user/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.3.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - remove @volontariapp/contracts
+
 ## 2.3.1
 
 ### Patch Changes